refactor: 타입 정의 파일 경로 수정 및 새로운 타입 추가
