+++ src/pages/explore/index.tsx
@@ -3,7 +3,7 @@ import { Settings2, LayoutGrid, List } from "lucide-react";
 import { Button } from "../../components/ui/button";
 import { FilteringModal } from "./_components/FilteringModal";
 import { EXPLORE_TABS, type ExploreTabValue } from "../../constants/explore";
-import type { Developer } from "../../types/developer";
+import type { Developer } from "../../types/developer.types";
 import { ListView } from "./_components/layout/ListView";
 import { GridView } from "./_components/layout/GridView";
 import { AppTabs } from "../../components/common/AppTabs";
