+++ src/pages/home/index.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { Bell, Plus, Search } from "lucide-react";
 import { useParams, useNavigate } from "react-router-dom";
-import type { Post } from "../../types/board";
+import type { Post } from "../../types/board.types";
 import { fetchPosts } from "../../lib/api";
 import { AppTabs } from "../../components/common/AppTabs";
 import { AddPostModal } from "./_components/modal/AddPostModal";
