+++ src/types/board.types.ts
